# perf: Improve scope information collection performance