--- a/packages/babel-traverse/src/path/context.ts
+++ b/packages/babel-traverse/src/path/context.ts
@@ -133,8 +133,31 @@ export function stop(this: NodePath) {
   this._traverseFlags |= SHOULD_SKIP | SHOULD_STOP;
 }
 
-export function setScope(this: NodePath, ignoreNoScope?: boolean) {
-  if (!ignoreNoScope && this.opts?.noScope) return;
+export function _forceSetScope(current: NodePath) {
+  let path = current.parentPath;
+
+  if (
+    // Skip method scope if is computed method key or decorator expression
+    ((current.key === "key" || current.listKey === "decorators") &&
+      path.isMethod()) ||
+    // Skip switch scope if for discriminant (`x` in `switch (x) {}`).
+    (current.key === "discriminant" && path.isSwitchStatement())
+  ) {
+    path = path.parentPath;
+  }
+
+  let target;
+  while (path && !target) {
+    target = path.scope;
+    path = path.parentPath;
+  }
+
+  current.scope = current.getScope(target);
+  current.scope?.init();
+}
+
+export function setScope(this: NodePath) {
+  if (this.opts?.noScope) return;
 
   let path = this.parentPath;
 
@@ -150,7 +173,7 @@ export function setScope(this: NodePath, ignoreNoScope?: boolean) {
 
   let target;
   while (path && !target) {
-    if (!ignoreNoScope && path.opts?.noScope) return;
+    if (path.opts?.noScope) return;
 
     target = path.scope;
     path = path.parentPath;