--- a/packages/babel-traverse/src/scope/index.ts
+++ b/packages/babel-traverse/src/scope/index.ts
@@ -1,6 +1,7 @@
 import Renamer from "./lib/renamer.ts";
 import type NodePath from "../path/index.ts";
 import traverse from "../index.ts";
+import traverseForScope from "./traverseForScope.ts";
 import Binding from "./binding.ts";
 import type { BindingKind } from "./binding.ts";
 import globals from "globals";
@@ -923,9 +924,6 @@ class Scope {
     // traverse does not visit the root node, here we explicitly collect
     // root node binding info when the root is not a Program.
     if (path.type !== "Program" && isExplodedVisitor(collectorVisitor)) {
-      for (const visit of collectorVisitor.enter) {
-        visit.call(state, path, state);
-      }
       const typeVisitors = collectorVisitor[path.type];
       if (typeVisitors) {
         for (const visit of typeVisitors.enter) {
@@ -933,7 +931,7 @@ class Scope {
         }
       }
     }
-    path.traverse(collectorVisitor, state);
+    traverseForScope(path, collectorVisitor, state);
     this.crawling = false;
 
     // register assignments