<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
Here we introduced a simplified traversal, which improved the overall performance by about 10%. (Well, I don’t understand why it brings such a big performance improvement. I tried more later but failed. 🤦‍♂️)

```
PS F:\babel\benchmark> node --expose-gc .\all\real-case-ts-mjs.mjs 
all/real-case-ts-mjs.mjs babel-parser-express.ts @ current: 36.76 ops/sec ±8.46% 30 runs (27ms)
all/real-case-ts-mjs.mjs babel-parser-express.ts @ baseline: 33.86 ops/sec ±5.73% 30 runs (30ms)
all/real-case-ts-mjs.mjs ts-parser.ts @ current: 6.51 ops/sec ±6.65% 30 runs (154ms)
all/real-case-ts-mjs.mjs ts-parser.ts @ baseline: 5.21 ops/sec ±4.09% 30 runs (192ms)
PS F:\babel\benchmark> node --expose-gc .\all\real-case-mjs-cjs.mjs
all/real-case-mjs-cjs.mjs ts-checker.mjs @ current: 1.11 ops/sec ±2.37% 30 runs (903ms)
all/real-case-mjs-cjs.mjs ts-checker.mjs @ baseline: 0.93 ops/sec ±2.46% 30 runs (1079ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ current: 8.29 ops/sec ±3.11% 30 runs (121ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ baseline: 7.36 ops/sec ±3.63% 30 runs (136ms)
```